+++ dbms/src/Debug/MockKVStore/MockRaftStoreProxy.cpp
@@ -601,7 +601,7 @@ std::tuple<RegionPtr, PrehandleResult> MockRaftStoreProxy::snapshot(
     uint64_t index,
     uint64_t term,
     std::optional<uint64_t> deadline_index,
-    bool cancel_after_prehandle)
+    std::optional<std::function<void()>> cancel_after_prehandle)
 {
     auto old_kv_region = kvs.getRegion(region_id);
     RUNTIME_CHECK(old_kv_region != nullptr);
@@ -628,7 +628,7 @@ std::tuple<RegionPtr, PrehandleResult> MockRaftStoreProxy::snapshot(
     uint64_t index,
     uint64_t term,
     std::optional<uint64_t> deadline_index,
-    bool cancel_after_prehandle)
+    std::optional<std::function<void()>> cancel_after_prehandle)
 {
     auto region = getRegion(region_id);
     RUNTIME_CHECK(region != nullptr);
@@ -665,6 +665,7 @@ std::tuple<RegionPtr, PrehandleResult> MockRaftStoreProxy::snapshot(
         auto rg = RegionPtrWithSnapshotFiles{new_kv_region, std::vector(prehandle_result.ingest_ids)};
         if (cancel_after_prehandle)
         {
+            cancel_after_prehandle.value()();
             kvs.releasePreHandledSnapshot(rg, tmt);
             return std::make_tuple(kvs.getRegion(region_id), prehandle_result);
         }
