+++ dbms/src/Storages/KVStore/FFI/ProxyFFI.cpp
@@ -666,6 +666,7 @@ RawCppPtr PreHandleSnapshot(
         CHECK_PARSE_PB_BUFF(region, region_buff.data, region_buff.len);
         auto & tmt = *server->tmt;
         auto & kvstore = tmt.getKVStore();
+        // TODO move this inside `preHandleSnapshotToFiles`, or change `genRegionPtr`.
         auto new_region = kvstore->genRegionPtr(std::move(region), peer_id, index, term);
         auto & region_table = tmt.getRegionTable();
         region_table.addPrehandlingRegion(*new_region);
